--- a/REQUIRE
+++ b/REQUIRE
@@ -1,6 +1,6 @@
 julia 0.5
 NullableArrays 0.1.0
-CategoricalArrays 0.0.6
+CategoricalArrays 0.1.2
 StatsBase 0.11.0
 GZip
 SortingAlgorithms