--- a/src/abstractdatatable/join.jl
+++ b/src/abstractdatatable/join.jl
@@ -15,76 +15,6 @@ similar_nullable{T,R}(dv::CategoricalArray{T,R}, dims::@compat(Union{Int, Tuple{
 similar_nullable(dt::AbstractDataTable, dims::Int) =
     DataTable(Any[similar_nullable(x, dims) for x in columns(dt)], copy(index(dt)))
 
-function join_idx(left, right, max_groups)
-    ## adapted from Wes McKinney's full_outer_join in pandas (file: src/join.pyx).
-
-    # NULL group in location 0
-
-    left_sorter, where, left_count = groupsort_indexer(left, max_groups)
-    right_sorter, where, right_count = groupsort_indexer(right, max_groups)
-
-    # First pass, determine size of result set
-    tcount = 0
-    rcount = 0
-    lcount = 0
-    for i in 1:(max_groups + 1)
-        lc = left_count[i]
-        rc = right_count[i]
-
-        if rc > 0 && lc > 0
-            tcount += lc * rc
-        elseif rc > 0
-            rcount += rc
-        else
-            lcount += lc
-        end
-    end
-
-    # group 0 is the NULL group
-    tposition = 0
-    lposition = 0
-    rposition = 0
-
-    left_pos = 0
-    right_pos = 0
-
-    left_indexer = Vector{Int}(tcount)
-    right_indexer = Vector{Int}(tcount)
-    leftonly_indexer = Vector{Int}(lcount)
-    rightonly_indexer = Vector{Int}(rcount)
-    for i in 1:(max_groups + 1)
-        lc = left_count[i]
-        rc = right_count[i]
-        if rc == 0
-            for j in 1:lc
-                leftonly_indexer[lposition + j] = left_pos + j
-            end
-            lposition += lc
-        elseif lc == 0
-            for j in 1:rc
-                rightonly_indexer[rposition + j] = right_pos + j
-            end
-            rposition += rc
-        else
-            for j in 1:lc
-                offset = tposition + (j-1) * rc
-                for k in 1:rc
-                    left_indexer[offset + k] = left_pos + j
-                    right_indexer[offset + k] = right_pos + k
-                end
-            end
-            tposition += lc * rc
-        end
-        left_pos += lc
-        right_pos += rc
-    end
-
-    ## (left_sorter, left_indexer, leftonly_indexer,
-    ##  right_sorter, right_indexer, rightonly_indexer)
-    (left_sorter[left_indexer], left_sorter[leftonly_indexer],
-     right_sorter[right_indexer], right_sorter[rightonly_indexer])
-end
-
 function sharepools{S,N}(v1::Union{CategoricalArray{S,N}, NullableCategoricalArray{S,N}},
                          v2::Union{CategoricalArray{S,N}, NullableCategoricalArray{S,N}},
                          index::Vector{S},
@@ -214,6 +144,199 @@ function sharepools(dt1::AbstractDataTable, dt2::AbstractDataTable)
     sharepools(refs1, refs2)
 end
 
+# helper structure for DataTables joining
+immutable _DataTableJoiner{DT1<:AbstractDataTable, DT2<:AbstractDataTable}
+    dtl::DT1
+    dtr::DT2
+    dtl_on::DT1
+    dtr_on::DT2
+    on_cols::Vector{Symbol}
+
+    function _DataTableJoiner(dtl::DT1, dtr::DT2, on::Union{Symbol,Vector{Symbol}})
+        on_cols = (isa(on, Symbol) ? fill(on::Symbol, 1) : on)::Vector{Symbol}
+        new(dtl, dtr, dtl[on_cols], dtr[on_cols], on_cols)
+    end
+end
+
+_DataTableJoiner{DT1<:AbstractDataTable, DT2<:AbstractDataTable}(dtl::DT1, dtr::DT2, on::Union{Symbol,Vector{Symbol}}) =
+    _DataTableJoiner{DT1,DT2}(dtl, dtr, on)
+
+# helper map between the row indices in original and joined frame
+immutable _RowIndexMap
+  orig::Vector{Int} # row indices in the original frame
+  join::Vector{Int} # row indices in the resulting joined frame
+end
+
+Base.length(x::_RowIndexMap) = length(x.orig)
+
+# fix the result of the rightjoin by taking the nonnull values from the right frame
+function fix_rightjoin_column!(res_col::AbstractArray, col_ix::Int, joiner::_DataTableJoiner,
+                              all_orig_left_ixs::Vector{Int}, rightonly_ixs::_RowIndexMap)
+    res_col[rightonly_ixs.join] = joiner.dtr_on[rightonly_ixs.orig, col_ix]
+    res_col
+end
+
+# since setindex!() for PoolDataArray is very slow,
+# it requires special handling
+function fix_rightjoin_column!{T,N}(res_col::Union{AbstractCategoricalArray{T,N}, AbstractNullableCategoricalArray{T,N}},
+                                    col_ix::Int, joiner::_DataTableJoiner,
+                                    all_orig_left_ixs::Vector{Int}, rightonly_ixs::_RowIndexMap)
+    left_col = joiner.dtl_on[col_ix]
+    right_col = joiner.dtr_on[col_ix]
+    if levels(left_col) == levels(right_col)
+        res_col.refs[rightonly_ixs.join] = right_col.refs[rightonly_ixs.orig]
+        return res_col
+    end
+    # merge the pools
+    # FIXME use sharedpools()?
+    newlevels, ordered = CategoricalArrays.mergelevels(levels(left_col), levels(right_col))
+    if length(newlevels) <= typemax(reftype(left_col))
+        newreftype = reftype(left_col)
+    elseif length(newlevels) <= typemax(reftype(right_col))
+        newreftype = reftype(right_col)
+    else
+        newreftype = reftype(length(newlevels))
+    end
+    new_refs = newreftype[
+            indexin(CategoricalArrays.index(left_col.pool), newlevels)[left_col.refs[all_orig_left_ixs]];
+            indexin(CategoricalArrays.index(right_col.pool), newlevels)[right_col.refs[rightonly_ixs.orig]]]
+    NullableCategoricalArray{T,N,newreftype}(new_refs, CategoricalPool(newlevels, ordered))
+end
+
+# composes the joined data frame using the maps between the left and right
+# frame rows and the indices of rows in the result
+function compose_joined_frame(joiner::_DataTableJoiner,
+                left_ixs::_RowIndexMap, leftonly_ixs::_RowIndexMap,
+                right_ixs::_RowIndexMap, rightonly_ixs::_RowIndexMap)
+    @assert length(left_ixs) == length(right_ixs)
+    # compose left half of the result taking all left columns
+    # FIXME is it still relevant? complicated way to do vcat that avoids expensive setindex!() for PooledDataVector
+    all_orig_left_ixs = [left_ixs.orig; leftonly_ixs.orig]
+    if length(leftonly_ixs) > 0
+        # permute the indices to restore left frame rows order
+        all_orig_left_ixs[[left_ixs.join; leftonly_ixs.join]] = all_orig_left_ixs
+    end
+    left_dt = DataTable(Any[resize!(col[all_orig_left_ixs], length(all_orig_left_ixs)+length(rightonly_ixs))
+                            for col in columns(joiner.dtl)],
+                        names(joiner.dtl))
+
+    # compose right half of the result taking all right columns excluding on
+    dtr_noon = without(joiner.dtr, joiner.on_cols)
+    # FIXME is it still relevant? complicated way to do vcat that avoids expensive setindex!() for PooledDataVector
+    # permutation to swap rightonly and leftonly rows
+    right_perm = [1:length(right_ixs);
+                  ((length(right_ixs)+length(rightonly_ixs)+1):
+                   (length(right_ixs)+length(rightonly_ixs)+length(leftonly_ixs)));
+                  ((length(right_ixs)+1):(length(right_ixs)+length(rightonly_ixs)))]
+    if length(leftonly_ixs) > 0
+        # compose right_perm with the permutation that restores left rows order
+        right_perm[[right_ixs.join; leftonly_ixs.join]] = right_perm[1:(length(right_ixs)+length(leftonly_ixs))]
+    end
+    all_orig_right_ixs = [right_ixs.orig; rightonly_ixs.orig]
+    right_dt = DataTable(Any[resize!(col[all_orig_right_ixs], length(all_orig_right_ixs)+length(leftonly_ixs))[right_perm]
+                             for col in columns(dtr_noon)],
+                         names(dtr_noon))
+    # merge left and right parts of the joined frame
+    res = hcat!(left_dt, right_dt)
+
+    if length(rightonly_ixs.join) > 0
+        # some left rows are nulls, so the values of the "on" columns
+        # need to be taken from the right
+        for (on_col_ix, on_col) in enumerate(joiner.on_cols)
+            res[on_col] = fix_rightjoin_column!(res[on_col], on_col_ix, joiner, all_orig_left_ixs, rightonly_ixs)
+        end
+    end
+    return res
+end
+
+# map the indices of the left and right joined frames
+# to the indices of the rows in the resulting frame
+# if `nothing` is given, the corresponding map is not built
+function update_row_maps!(left_frame::AbstractDataTable, right_frame::AbstractDataTable,
+                    right_dict::RowGroupDict,
+                    left_ixs::Union{Void, _RowIndexMap},
+                    leftonly_ixs::Union{Void, _RowIndexMap},
+                    right_ixs::Union{Void, _RowIndexMap},
+                    rightonly_mask::Union{Void, Vector{Bool}})
+    # helper functions
+    update!(ixs::Void, orig_ix::Int, join_ix::Int, count::Int = 1) = ixs
+    function update!(ixs::_RowIndexMap, orig_ix::Int, join_ix::Int, count::Int = 1)
+        for i in 1:count
+            push!(ixs.orig, orig_ix)
+        end
+        for i in join_ix:(join_ix+count-1)
+            push!(ixs.join, i)
+        end
+        ixs
+    end
+    update!(ixs::Void, orig_ixs::AbstractArray, join_ix::Int) = ixs
+    function update!(ixs::_RowIndexMap, orig_ixs::AbstractArray, join_ix::Int)
+        append!(ixs.orig, orig_ixs)
+        for i in join_ix:(join_ix+length(orig_ixs)-1)
+            push!(ixs.join, i)
+        end
+        ixs
+    end
+    update!(ixs::Void, orig_ixs::AbstractArray) = ixs
+    update!(mask::Vector{Bool}, orig_ixs::AbstractArray) = (mask[orig_ixs] = false)
+
+    # iterate over left rows and compose the left<->right index map
+    next_join_ix = 1
+    for l_ix in 1:nrow(left_frame)
+        r_ixs = get(right_dict, left_frame, l_ix)
+        if isempty(r_ixs)
+            update!(leftonly_ixs, l_ix, next_join_ix)
+            next_join_ix += 1
+        else
+            update!(left_ixs, l_ix, next_join_ix, length(r_ixs))
+            update!(right_ixs, r_ixs, next_join_ix)
+            update!(rightonly_mask, r_ixs)
+            next_join_ix += length(r_ixs)
+        end
+    end
+end
+
+# map the row indices of the left and right joined frames
+# to the indices of rows in the resulting frame
+# returns the 4-tuple of row indices maps for
+# - matching left rows
+# - non-matching left rows
+# - matching right rows
+# - non-matching right rows
+# if false is provided, the corresponding map is not built and the
+# tuple element is empty _RowIndexMap
+function update_row_maps!(
+    left_frame::AbstractDataTable, right_frame::AbstractDataTable,
+    right_dict::RowGroupDict,
+    map_left::Bool, map_leftonly::Bool,
+    map_right::Bool, map_rightonly::Bool)
+    init_map(dt::AbstractDataTable, init::Bool) = init ?
+        _RowIndexMap(sizehint!(Vector{Int}(), nrow(dt)),
+                     sizehint!(Vector{Int}(), nrow(dt))) :
+         nothing
+    to_bimap(x::_RowIndexMap) = x
+    to_bimap(::Void) = _RowIndexMap(Vector{Int}(), Vector{Int}())
+
+    # init maps as requested
+    left_ixs = init_map(left_frame, map_left)
+    leftonly_ixs = init_map(left_frame, map_leftonly)
+    right_ixs = init_map(right_frame, map_right)
+    rightonly_mask = map_rightonly ? fill(true, nrow(right_frame)) : nothing
+    update_row_maps!(left_frame, right_frame, right_dict,
+                     left_ixs, leftonly_ixs, right_ixs, rightonly_mask)
+    if map_rightonly
+        rightonly_orig_ixs = (1:length(rightonly_mask))[rightonly_mask]
+        rightonly_ixs = _RowIndexMap(rightonly_orig_ixs,
+                         collect(length(right_ixs.orig)+
+                                 (leftonly_ixs === nothing ? 0 : length(leftonly_ixs))+
+                                 (1:length(rightonly_orig_ixs))))
+    else
+        rightonly_ixs = nothing
+    end
+
+    return to_bimap(left_ixs), to_bimap(leftonly_ixs),
+           to_bimap(right_ixs), to_bimap(rightonly_ixs)
+end
 
 """
 Join two DataTables
@@ -270,61 +393,61 @@ join(name, job, kind = :cross)
 """
 function Base.join(dt1::AbstractDataTable,
                    dt2::AbstractDataTable;
-                   on::@compat(Union{Symbol, Vector{Symbol}}) = Symbol[],
+                   on::Union{Symbol, Vector{Symbol}} = Symbol[],
                    kind::Symbol = :inner)
     if kind == :cross
-        if on != Symbol[]
-            throw(ArgumentError("Cross joins don't use argument 'on'."))
-        end
+        (on == Symbol[]) || throw(ArgumentError("Cross joins don't use argument 'on'."))
         return crossjoin(dt1, dt2)
     elseif on == Symbol[]
         throw(ArgumentError("Missing join argument 'on'."))
     end
 
-    dv1, dv2 = sharepools(dt1[on], dt2[on])
-
-    left_idx, leftonly_idx, right_idx, rightonly_idx =
-        join_idx(dv1.refs, dv2.refs, length(dv1.pool))
+    joiner = _DataTableJoiner(dt1, dt2, on)
 
     if kind == :inner
-        dt2w = without(dt2, on)
-
-        left = dt1[left_idx, :]
-        right = dt2w[right_idx, :]
-
-        return hcat!(left, right)
+        compose_joined_frame(joiner, update_row_maps!(joiner.dtl_on, joiner.dtr_on,
+                             group_rows(joiner.dtr_on),
+                             true, false, true, false)...)
     elseif kind == :left
-        dt2w = without(dt2, on)
-
-        left = dt1[[left_idx; leftonly_idx], :]
-        right = vcat(dt2w[right_idx, :],
-                     similar_nullable(dt2w, length(leftonly_idx)))
-
-        return hcat!(left, right)
+        compose_joined_frame(joiner, update_row_maps!(joiner.dtl_on, joiner.dtr_on,
+                             group_rows(joiner.dtr_on),
+                             true, true, true, false)...)
     elseif kind == :right
-        dt1w = without(dt1, on)
-
-        left = vcat(dt1w[left_idx, :],
-                    similar_nullable(dt1w, length(rightonly_idx)))
-        right = dt2[[right_idx; rightonly_idx], :]
-
-        return hcat!(left, right)
+        right_ixs, rightonly_ixs, left_ixs, leftonly_ixs =
+            update_row_maps!(joiner.dtr_on, joiner.dtl_on,
+                            group_rows(joiner.dtl_on),
+                            true, true, true, false)
+        compose_joined_frame(joiner, left_ixs, leftonly_ixs, right_ixs, rightonly_ixs)
     elseif kind == :outer
-        dt1w, dt2w = without(dt1, on),  without(dt2, on)
-
-        mixed = hcat!(dt1[left_idx, :], dt2w[right_idx, :])
-        leftonly = hcat!(dt1[leftonly_idx, :],
-                         similar_nullable(dt2w, length(leftonly_idx)))
-        rightonly = hcat!(similar_nullable(dt1w, length(rightonly_idx)),
-                          dt2[rightonly_idx, :])
-
-        return vcat(mixed, leftonly, rightonly)
+        compose_joined_frame(joiner, update_row_maps!(joiner.dtl_on, joiner.dtr_on,
+                             group_rows(joiner.dtr_on),
+                             true, true, true, true)...)
     elseif kind == :semi
-        dt1[unique(left_idx), :]
+        # hash the right rows
+        dtr_on_grp = group_rows(joiner.dtr_on)
+        # iterate over left rows and leave those found in right
+        left_ixs = Vector{Int}()
+        sizehint!(left_ixs, nrow(joiner.dtl))
+        for l_ix in 1:nrow(joiner.dtl_on)
+            if in(dtr_on_grp, joiner.dtl_on, l_ix)
+                push!(left_ixs, l_ix)
+            end
+        end
+        return joiner.dtl[left_ixs, :]
     elseif kind == :anti
-        dt1[leftonly_idx, :]
+        # hash the right rows
+        dtr_on_grp = group_rows(joiner.dtr_on)
+        # iterate over left rows and leave those not found in right
+        leftonly_ixs = Vector{Int}()
+        sizehint!(leftonly_ixs, nrow(joiner.dtl))
+        for l_ix in 1:nrow(joiner.dtl_on)
+            if !in(dtr_on_grp, joiner.dtl_on, l_ix)
+                push!(leftonly_ixs, l_ix)
+            end
+        end
+        return joiner.dtl[leftonly_ixs, :]
     else
-        throw(ArgumentError("Unknown kind of join requested"))
+        throw(ArgumentError("Unknown kind ($kind) of join requested"))
     end
 end