The implementation @alyst wrote for https://github.com/JuliaStats/DataFrames.jl/pull/850 is very well done. It's clearly more performant than the current master and what I wrote for https://github.com/JuliaData/DataTables.jl/pull/3. I merged https://github.com/JuliaStats/DataFrames.jl/pull/850 with DataTables but if I could get some keen eyes, particularly @alyst's to make sure I didn't make any meaningful errors while fixing merge conflicts that would be appreciated. I merged benchmarks from https://github.com/JuliaStats/DataFrames.jl/pull/850, https://github.com/JuliaData/DataTables.jl/pull/3, and https://github.com/JuliaData/DataTables.jl/pull/12 to generate this list of benchmarks.

**benchmarks**
```julia
using DataTables
using BenchmarkTools
using Distributions

srand(1);
a = rand([1, 2, Nullable(), 4], 20);
b = rand([:a, :b], 20);
c = rand(["a", "b"], 20);
d = rand(rand(2), 20);
small = DataTable(A = a, B = b, C = c, D = c);
A = repeat(rand(8891), inner=2);
B = repeat(rand(8891), inner=2);
C = repeat(rand(8891), inner=2);
D = repeat(rand(8891), inner=2);
E = repeat(rand(8891), inner=2);
F = repeat(rand(17782));
large = DataTable(A = A, B = B, C = C, D = D, E = E, F = F);
small_ordered = DataTable(v1 = categorical(collect(1:1000)), v2 = categorical(fill(1, 1000)));
small_unordered = DataTable(v1 = categorical(rand(1000)), v2 = categorical(rand(1000)));
large_ordered = DataTable(v1 = categorical(collect(1:100000)), v2 = categorical(fill(1, 100000)));
large_unordered = DataTable(v1 = categorical(rand(100000)), v2 = categorical(rand(100000)));
wide_ordered = DataTable(v1 = categorical(collect(1:100000)),
                         v2 = categorical(fill(1, 100000)),
                         v3 = categorical(collect(1:100000)),
                         v4 = categorical(collect(1:100000)));

dt1 = DataTable(v1 = CategoricalArray(repeat(1:10, inner=1000)), v2 = CategoricalArray(repeat(1:10, outer=1000)));
dt2 = DataTable(v1 = CategoricalArray(repeat(1:100, inner=100)), v2 = CategoricalArray(repeat(1:100, outer=100)));
dt3 = hcat(dt1, dt2);

function random_frame(nrow::Int, col_values::Dict{Symbol, Any})
  DataTable(Any[isa(col_values[key], CategoricalArray) ?
                sample(col_values[key], nrow) :
                sample(col_values[key], nrow) for key in keys(col_values)],
            keys(col_values) |> collect)
end

function random_join(kind::Symbol, nrow_left::Int, nrow_right::Int,
                     on_col_values::Dict{Symbol, Any},
                     left_col_values::Dict{Symbol, Any},
                     right_col_values::Dict{Symbol, Any})
  dtl = random_frame(nrow_left, merge(on_col_values, left_col_values))
  dtr = random_frame(nrow_right, merge(on_col_values, right_col_values))
  join(dtl, dtr, on = keys(on_col_values) |> collect, kind = kind)
end

function f(n::Int)
    for i in 1:n
    r = random_join(:outer, 1000, 2000,
                Dict{Symbol,Any}(:A => 1:10, :B => [:A, :B, :C, :D],
                                 :C => 1:10, :D => 1:10),
                Dict{Symbol,Any}(:E => 1:10, :F => [:A, :B, :C, :D]),
                Dict{Symbol,Any}(:G => 1:10, :H => [:A, :B, :C, :D]))
    end
end
f(1)

function h(n::Int)
    for i in 1:n
        r = random_join(:outer, 10000, 20000,
                       Dict{Symbol,Any}(:A => collect(1:10000)),
                       Dict{Symbol,Any}(:B => collect(1:10000)),
                       Dict{Symbol,Any}(:C => collect(1:10000)))
        end
    end

h(1)

@benchmark groupby(small, [:A, :B])
@benchmark groupby(large, [:A, :B])
@benchmark groupby(large, [:A, :B, :C, :D, :E])
@benchmark groupby(small_ordered, [:v1, :v2])
@benchmark groupby(small_unordered, [:v1, :v2])
@benchmark groupby(large_ordered, [:v1, :v2])
@benchmark groupby(large_unordered, [:v1, :v2])
@benchmark groupby(wide_ordered, [:v1, :v2, :v3, :v4])

@benchmark groupby(dt1, [:v1, :v2])
@benchmark groupby(dt2, [:v1, :v2])
@benchmark groupby(dt3, [:v1, :v2, :v1_1, :v2_1])

@time f(100)
@benchmark f(100)
@time h(100)
@benchmark h(100)
```

results
```julia
julia> @benchmark groupby(small, [:A, :B])
  #master
  memory estimate:  11.95 KiB
  allocs estimate:  185
  median time:      51.035 μs (0.00% GC)
  #3
  memory estimate:  15.91 KiB
  allocs estimate:  516
  median time:      21.434 μs (0.00% GC)
  #this pr/DataFrames:850
  memory estimate:  6.55 KiB
  allocs estimate:  130
  median time:      4.213 μs (0.00% GC)


julia> @benchmark groupby(large, [:A, :B])
  #master
  memory estimate:  1.26 GiB
  allocs estimate:  388362
  median time:      1.418 s (21.80% GC)
  #3
  memory estimate:  18.87 MiB
  allocs estimate:  815974
  median time:      32.828 ms (14.98% GC)
  #this pr/DataFrames:850
  memory estimate:  1.39 MiB
  allocs estimate:  34592
  median time:      1.648 ms (0.00% GC)

julia> @benchmark groupby(large, [:A, :B, :C, :D, :E])
  #master
  ERROR: InexactError()
  #3
  memory estimate:  39.59 MiB
  allocs estimate:  1834418
  median time:      63.197 ms (16.02% GC)
  #this pr/DataFrames:850
  memory estimate:  2.18 MiB
  allocs estimate:  86411
  median time:      3.514 ms (0.00% GC)

julia> @benchmark groupby(small_ordered, [:v1, :v2])
  #master
  memory estimate:  329.38 KiB
  allocs estimate:  9412
  median time:      2.076 ms (0.00% GC)
  #3
  memory estimate:  923.17 KiB
  allocs estimate:  34720
  median time:      1.237 ms (0.00% GC)
  #this pr/DataFrames:850
  memory estimate:  57.98 KiB
  allocs estimate:  43
  median time:      51.100 μs (0.00% GC)

julia> @benchmark groupby(small_unordered, [:v1, :v2])
  #master
  memory estimate:  16.72 MiB
  allocs estimate:  15905
  median time:      14.618 ms (0.00% GC)
  #3
  memory estimate:  1018.45 KiB
  allocs estimate:  38912
  median time:      1.526 ms (0.00% GC)
  #this pr/DataFrames:850
  memory estimate:  57.98 KiB
  allocs estimate:  43
  median time:      56.681 μs (0.00% GC)

julia> @benchmark groupby(large_ordered, [:v1, :v2])
  #master
  memory estimate:  44.40 MiB
  allocs estimate:  1890360
  median time:      216.415 ms (0.00% GC)
  #3
  memory estimate:  104.09 MiB
  allocs estimate:  4357825
  median time:      313.996 ms (24.12% GC)
  #this pr/DataFrames:850
  memory estimate:  5.58 MiB
  allocs estimate:  50
  median time:      4.952 ms (0.00% GC)

julia> @benchmark groupby(large_unordered, [:v1, :v2])
  #master
  ERROR: InexactError()
  #3
  memory estimate:  104.21 MiB
  allocs estimate:  4362897
  median time:      312.446 ms (22.47% GC)
  #this pr/DataFrames:850
  memory estimate:  5.58 MiB
  allocs estimate:  50
  median time:      5.192 ms (0.00% GC)

julia> @benchmark groupby(wide_ordered, [:v1, :v2, :v3, :v4])
  #master
  ERROR: InexactError()
  #3
  memory estimate:  175.00 MiB
  allocs estimate:  7968344
  median time:      484.698 ms (22.13% GC)
  #this pr/DataFrames:850
  memory estimate:  5.58 MiB
  allocs estimate:  54
  median time:      5.723 ms (0.00% GC)

julia> @benchmark groupby(dt1, [:v1, :v2])
  #master
  memory estimate:  1.22 MiB
  allocs estimate:  66536
  median time:      3.547 ms (0.00% GC)
  #3
  memory estimate:  10.68 MiB
  allocs estimate:  476970
  median time:      19.039 ms (0.00% GC)
  #this pr/DataFrames:850
  memory estimate:  3.33 MiB
  allocs estimate:  115865
  median time:      2.119 ms (0.00% GC)

julia> @benchmark groupby(dt2, [:v1, :v2])
  #master
  memory estimate:  1.99 MiB
  allocs estimate:  85972
  median time:      4.067 ms (0.00% GC)
  #3
  memory estimate:  10.06 MiB
  allocs estimate:  409364
  median time:      18.255 ms (0.00% GC)
  #this pr/DataFrames:850
  memory estimate:  599.53 KiB
  allocs estimate:  50
  median time:      332.527 μs (0.00% GC)

julia> @benchmark groupby(dt3, [:v1, :v2, :v1_1, :v2_1])
  #master
  memory estimate:  20.45 MiB
  allocs estimate:  238751
  median time:      18.464 ms (13.56% GC)
  #3
  memory estimate:  16.74 MiB
  allocs estimate:  744976
  median time:      32.042 ms (15.40% GC)
  #this pr/DataFrames:850
  memory estimate:  599.70 KiB
  allocs estimate:  54
  median time:      416.803 μs (0.00% GC)

julia> @time f(100)
  #master
  0.519709 seconds (5.89 M allocations: 268.796 MB, 8.70% gc time)
  #3
  0.523668 seconds (5.89 M allocations: 268.796 MB, 7.42% gc time)
  #this pr/DataFrames:850
  0.487742 seconds (5.89 M allocations: 268.796 MB, 7.94% gc time)

julia> @benchmark f(100)
  #master
  memory estimate:  268.61 MiB
  allocs estimate:  5888594
  median time:      483.714 ms (7.89% GC)
  #3
  memory estimate:  268.61 MiB
  allocs estimate:  5888594
  median time:      524.871 ms (8.02% GC)
  #this pr/DataFrames:850
  memory estimate:  268.61 MiB
  allocs estimate:  5888594
  median time:      508.119 ms (7.48% GC)

julia> @time h(100)
  #master
  1.288805 seconds (9.32 M allocations: 733.918 MB, 16.57% gc time)
  #3
  1.230394 seconds (9.31 M allocations: 733.837 MB, 8.34% gc time)
  #this pr/DataFrames:850
  1.220218 seconds (9.31 M allocations: 733.837 MB, 9.96% gc time)

julia> @benchmark h(100)
  #master
  memory estimate:  733.79 MiB
  allocs estimate:  9312901
  median time:      1.277 s (9.13% GC)
  #3
  memory estimate:  733.81 MiB
  allocs estimate:  9312588
  median time:      1.275 s (8.45% GC)
  #this pr/DataFrames:850
  memory estimate:  733.79 MiB
  allocs estimate:  9312588
  median time:      1.186 s (9.87% GC)
```