@cjprybol 2017-02-22
Yes, thanks for mentioning those (joins & `nonunique()`), let's not leave anything behind! `nonunique()` was easy to add but I couldn't get the joins to work

```julia
julia> dt1 = DataTable(a = shuffle!([1:10;]),
                       b = [:A,:B][rand(1:2, 10)],
                       v1 = randn(10))
10×3 DataTables.DataTable
│ Row │ a  │ b  │ v1         │
├─────┼────┼────┼────────────┤
│ 1   │ 9  │ :A │ -1.72976   │
│ 2   │ 5  │ :A │ 0.795949   │
│ 3   │ 1  │ :A │ 0.670062   │
│ 4   │ 7  │ :B │ 0.550852   │
│ 5   │ 4  │ :A │ -0.0633746 │
│ 6   │ 10 │ :B │ 1.33694    │
│ 7   │ 6  │ :B │ -0.0731486 │
│ 8   │ 2  │ :B │ -0.745464  │
│ 9   │ 8  │ :A │ -1.22006   │
│ 10  │ 3  │ :B │ -0.0531773 │

julia> dt2 = DataTable(a = shuffle!(reverse([1:5;])),
                       b2 = [:A,:B,:C][rand(1:3, 5)],
                       v2 = randn(5))
5×3 DataTables.DataTable
│ Row │ a │ b2 │ v2        │
├─────┼───┼────┼───────────┤
│ 1   │ 1 │ :A │ 0.0704676 │
│ 2   │ 5 │ :A │ 0.341794  │
│ 3   │ 2 │ :C │ 1.73517   │
│ 4   │ 3 │ :A │ 1.29992   │
│ 5   │ 4 │ :B │ 0.206364  │

julia> m1 = join(dt1, dt2, on = :a)
0×5 DataTables.DataTable
```

I merged your join.jl as seen [here](https://gist.github.com/cjprybol/3ee3bacac66098672a522342d1362e4b). Any idea what I'm missing?