@alyst 2017-02-22
@cjprybol Strange. Could you please give me the link to the Git branch of DataTables with joins merged so that I can try debugging it?