@cjprybol 2017-03-06
**comparison code**
```julia
using DataTables
using BenchmarkTools
using Distributions

srand(1);
a = rand([1, 2, Nullable(), 4], 20);
b = rand([:a, :b], 20);
c = rand(["a", "b"], 20);
d = rand(rand(2), 20);
small = DataTable(A = a, B = b, C = c, D = c);
A = repeat(rand(8891), inner=2);
B = repeat(rand(8891), inner=2);
C = repeat(rand(8891), inner=2);
D = repeat(rand(8891), inner=2);
E = repeat(rand(8891), inner=2);
F = repeat(rand(17782));
large = DataTable(A = A, B = B, C = C, D = D, E = E, F = F);
small_ordered = DataTable(v1 = categorical(collect(1:1000)), v2 = categorical(fill(1, 1000)));
small_unordered = DataTable(v1 = categorical(rand(1000)), v2 = categorical(rand(1000)));
large_ordered = DataTable(v1 = categorical(collect(1:100000)), v2 = categorical(fill(1, 100000)));
large_unordered = DataTable(v1 = categorical(rand(100000)), v2 = categorical(rand(100000)));
wide_ordered = DataTable(v1 = categorical(collect(1:100000)),
                         v2 = categorical(fill(1, 100000)),
                         v3 = categorical(collect(1:100000)),
                         v4 = categorical(collect(1:100000)));

dt1 = DataTable(v1 = CategoricalArray(repeat(1:10, inner=1000)), v2 = CategoricalArray(repeat(1:10, outer=1000)));
dt2 = DataTable(v1 = CategoricalArray(repeat(1:100, inner=100)), v2 = CategoricalArray(repeat(1:100, outer=100)));
dt3 = hcat(dt1, dt2);

function random_frame(nrow::Int, col_values::Dict{Symbol, Any})
  DataTable(Any[isa(col_values[key], CategoricalArray) ?
                sample(col_values[key], nrow) :
                sample(col_values[key], nrow) for key in keys(col_values)],
            collect(keys(col_values)))
end

function random_join(kind::Symbol, nrow_left::Int, nrow_right::Int,
                     on_col_values::Dict{Symbol, Any},
                     left_col_values::Dict{Symbol, Any},
                     right_col_values::Dict{Symbol, Any})
  dtl = random_frame(nrow_left, merge(on_col_values, left_col_values))
  dtr = random_frame(nrow_right, merge(on_col_values, right_col_values))
  join(dtl, dtr, on = collect(keys(on_col_values)), kind = kind)
end

function h(n::Int)
    for i in 1:n
        r = random_join(:outer, 10000, 20000,
                       Dict{Symbol,Any}(:A => collect(1:10000)),
                       Dict{Symbol,Any}(:B => collect(1:10000)),
                       Dict{Symbol,Any}(:C => collect(1:10000)))
        end
    end

h(1)

@benchmark groupby(small, [:A, :B])
@benchmark groupby(large, [:A, :B])
@benchmark groupby(large, [:A, :B, :C, :D, :E])
@benchmark groupby(small_ordered, [:v1, :v2])
@benchmark groupby(small_unordered, [:v1, :v2])
@benchmark groupby(large_ordered, [:v1, :v2])
@benchmark groupby(large_unordered, [:v1, :v2])
@benchmark groupby(wide_ordered, [:v1, :v2, :v3, :v4])

@benchmark groupby(dt1, [:v1, :v2])
@benchmark groupby(dt2, [:v1, :v2])
@benchmark groupby(dt3, [:v1, :v2, :v1_1, :v2_1])
@benchmark h(100)
```

**results**
```julia
julia> @benchmark groupby(small, [:A, :B])
# master
BenchmarkTools.Trial:
  memory estimate:  11.95 KiB
  allocs estimate:  185
  --------------
  minimum time:     39.277 μs (0.00% GC)
  median time:      45.597 μs (0.00% GC)
  mean time:        50.339 μs (3.90% GC)
  maximum time:     4.514 ms (97.81% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  5.73 KiB
  allocs estimate:  104
  --------------
  minimum time:     3.132 μs (0.00% GC)
  median time:      3.681 μs (0.00% GC)
  mean time:        4.895 μs (18.08% GC)
  maximum time:     493.746 μs (98.02% GC)
  --------------
  samples:          10000
  evals/sample:     9
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(large, [:A, :B])
# master
BenchmarkTools.Trial:
  memory estimate:  1.26 GiB
  allocs estimate:  388362
  --------------
  minimum time:     1.256 s (17.89% GC)
  median time:      1.373 s (23.61% GC)
  mean time:        1.393 s (23.28% GC)
  maximum time:     1.570 s (27.03% GC)
  --------------
  samples:          4
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  1.39 MiB
  allocs estimate:  34592
  --------------
  minimum time:     1.281 ms (0.00% GC)
  median time:      1.694 ms (0.00% GC)
  mean time:        1.884 ms (7.23% GC)
  maximum time:     6.244 ms (50.07% GC)
  --------------
  samples:          2634
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(large, [:A, :B, :C, :D, :E])
# master
ERROR: InexactError()
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at ./array.jl:415
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at /Users/Cameron/julia/usr/lib/julia/sys.dylib:?
 in groupby(::DataTables.DataTable, ::Array{Symbol,1}) at /Users/Cameron/.julia/v0.5/DataTables/src/groupeddatatable/grouping.jl:142
 in ##core#281() at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:283
 in ##sample#282(::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:289
 in #_run#5(::Bool, ::String, ::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:317
 in (::BenchmarkTools.#kw##_run)(::Array{Any,1}, ::BenchmarkTools.#_run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in anonymous at ./<missing>:?
 in #run_result#16(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:33
 in (::BenchmarkTools.#kw##run_result)(::Array{Any,1}, ::BenchmarkTools.#run_result, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in #run#17(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:36
 in (::Base.#kw##run)(::Array{Any,1}, ::Base.#run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#280")}) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:71
# this PR
BenchmarkTools.Trial:
  memory estimate:  2.18 MiB
  allocs estimate:  86411
  --------------
  minimum time:     2.739 ms (0.00% GC)
  median time:      3.131 ms (0.00% GC)
  mean time:        3.376 ms (4.99% GC)
  maximum time:     7.522 ms (36.13% GC)
  --------------
  samples:          1475
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(small_ordered, [:v1, :v2])
# master
BenchmarkTools.Trial:
  memory estimate:  329.38 KiB
  allocs estimate:  9412
  --------------
  minimum time:     1.832 ms (0.00% GC)
  median time:      2.038 ms (0.00% GC)
  mean time:        2.191 ms (2.20% GC)
  maximum time:     7.613 ms (59.39% GC)
  --------------
  samples:          2272
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  57.98 KiB
  allocs estimate:  43
  --------------
  minimum time:     47.216 μs (0.00% GC)
  median time:      51.638 μs (0.00% GC)
  mean time:        66.316 μs (11.67% GC)
  maximum time:     3.995 ms (97.59% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(small_unordered, [:v1, :v2])
# master
BenchmarkTools.Trial:
  memory estimate:  16.72 MiB
  allocs estimate:  15905
  --------------
  minimum time:     10.192 ms (0.00% GC)
  median time:      12.566 ms (17.11% GC)
  mean time:        12.544 ms (13.95% GC)
  maximum time:     16.632 ms (21.00% GC)
  --------------
  samples:          398
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  57.98 KiB
  allocs estimate:  43
  --------------
  minimum time:     50.458 μs (0.00% GC)
  median time:      55.520 μs (0.00% GC)
  mean time:        70.863 μs (11.01% GC)
  maximum time:     6.109 ms (85.02% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(large_ordered, [:v1, :v2])
# master
BenchmarkTools.Trial:
  memory estimate:  44.40 MiB
  allocs estimate:  1890360
  --------------
  minimum time:     214.346 ms (6.45% GC)
  median time:      223.379 ms (6.72% GC)
  mean time:        222.742 ms (6.97% GC)
  maximum time:     236.645 ms (7.14% GC)
  --------------
  samples:          23
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  5.58 MiB
  allocs estimate:  50
  --------------
  minimum time:     3.262 ms (0.00% GC)
  median time:      4.867 ms (0.00% GC)
  mean time:        5.118 ms (7.04% GC)
  maximum time:     37.738 ms (5.83% GC)
  --------------
  samples:          972
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(large_unordered, [:v1, :v2])
# master
ERROR: InexactError()
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at ./array.jl:415
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at /Users/Cameron/julia/usr/lib/julia/sys.dylib:?
 in groupby(::DataTables.DataTable, ::Array{Symbol,1}) at /Users/Cameron/.julia/v0.5/DataTables/src/groupeddatatable/grouping.jl:142
 in ##core#293() at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:283
 in ##sample#294(::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:289
 in #_run#9(::Bool, ::String, ::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:317
 in (::BenchmarkTools.#kw##_run)(::Array{Any,1}, ::BenchmarkTools.#_run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in anonymous at ./<missing>:?
 in #run_result#16(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:33
 in (::BenchmarkTools.#kw##run_result)(::Array{Any,1}, ::BenchmarkTools.#run_result, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in #run#17(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:36
 in (::Base.#kw##run)(::Array{Any,1}, ::Base.#run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#292")}) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:71
# this PR
BenchmarkTools.Trial:
  memory estimate:  5.58 MiB
  allocs estimate:  50
  --------------
  minimum time:     3.482 ms (0.00% GC)
  median time:      5.201 ms (0.00% GC)
  mean time:        5.302 ms (6.70% GC)
  maximum time:     10.012 ms (16.90% GC)
  --------------
  samples:          939
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(wide_ordered, [:v1, :v2, :v3, :v4])
# master
ERROR: InexactError()
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at ./array.jl:415
 in setindex!(::Array{UInt32,1}, ::Int64, ::Int64) at /Users/Cameron/julia/usr/lib/julia/sys.dylib:?
 in groupby(::DataTables.DataTable, ::Array{Symbol,1}) at /Users/Cameron/.julia/v0.5/DataTables/src/groupeddatatable/grouping.jl:142
 in ##core#296() at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:283
 in ##sample#297(::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:289
 in #_run#10(::Bool, ::String, ::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:317
 in (::BenchmarkTools.#kw##_run)(::Array{Any,1}, ::BenchmarkTools.#_run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in anonymous at ./<missing>:?
 in #run_result#16(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:33
 in (::BenchmarkTools.#kw##run_result)(::Array{Any,1}, ::BenchmarkTools.#run_result, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in #run#17(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:36
 in (::Base.#kw##run)(::Array{Any,1}, ::Base.#run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#295")}) at /Users/Cameron/.julia/v0.5/BenchmarkTools/src/execution.jl:71
# this PR
BenchmarkTools.Trial:
  memory estimate:  5.58 MiB
  allocs estimate:  54
  --------------
  minimum time:     4.141 ms (0.00% GC)
  median time:      5.545 ms (0.00% GC)
  mean time:        5.843 ms (5.73% GC)
  maximum time:     11.428 ms (17.77% GC)
  --------------
  samples:          852
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(dt1, [:v1, :v2])
# master
BenchmarkTools.Trial:
  memory estimate:  1.22 MiB
  allocs estimate:  66536
  --------------
  minimum time:     3.511 ms (0.00% GC)
  median time:      3.882 ms (0.00% GC)
  mean time:        4.139 ms (2.13% GC)
  maximum time:     11.563 ms (20.03% GC)
  --------------
  samples:          1203
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  2.13 MiB
  allocs estimate:  76265
  --------------
  minimum time:     1.334 ms (0.00% GC)
  median time:      1.421 ms (0.00% GC)
  mean time:        1.832 ms (19.33% GC)
  maximum time:     8.033 ms (64.45% GC)
  --------------
  samples:          2716
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(dt2, [:v1, :v2])
# master
BenchmarkTools.Trial:
  memory estimate:  1.99 MiB
  allocs estimate:  85972
  --------------
  minimum time:     4.114 ms (0.00% GC)
  median time:      4.627 ms (0.00% GC)
  mean time:        4.868 ms (3.54% GC)
  maximum time:     8.601 ms (28.77% GC)
  --------------
  samples:          1023
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  599.53 KiB
  allocs estimate:  50
  --------------
  minimum time:     247.364 μs (0.00% GC)
  median time:      314.495 μs (0.00% GC)
  mean time:        385.010 μs (12.40% GC)
  maximum time:     4.513 ms (86.66% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark groupby(dt3, [:v1, :v2, :v1_1, :v2_1])
# master
BenchmarkTools.Trial:
  memory estimate:  20.45 MiB
  allocs estimate:  238751
  --------------
  minimum time:     15.945 ms (0.00% GC)
  median time:      20.444 ms (12.86% GC)
  mean time:        21.459 ms (12.40% GC)
  maximum time:     40.995 ms (11.00% GC)
  --------------
  samples:          233
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  599.70 KiB
  allocs estimate:  54
  --------------
  minimum time:     339.325 μs (0.00% GC)
  median time:      420.928 μs (0.00% GC)
  mean time:        514.197 μs (10.94% GC)
  maximum time:     4.273 ms (74.66% GC)
  --------------
  samples:          9584
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark h(100)
# master
BenchmarkTools.Trial:
  memory estimate:  733.79 MiB
  allocs estimate:  9311593
  --------------
  minimum time:     1.215 s (10.20% GC)
  median time:      1.249 s (10.20% GC)
  mean time:        1.241 s (10.32% GC)
  maximum time:     1.256 s (10.34% GC)
  --------------
  samples:          5
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
# this PR
BenchmarkTools.Trial:
  memory estimate:  1.33 GiB
  allocs estimate:  40137024
  --------------
  minimum time:     3.107 s (6.17% GC)
  median time:      3.159 s (6.13% GC)
  mean time:        3.159 s (6.13% GC)
  maximum time:     3.211 s (6.09% GC)
  --------------
  samples:          2
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

the `groupby(dt1, [:v1, :v2])` improved since we started, the random join benchmark @alyst wrote looks like it is now doing worse than master in contrast to the significant improvements he saw when designing it against `PooledDataArray` https://github.com/JuliaStats/DataFrames.jl/pull/850#issue-99662572.

Plan forward:
- [ ] merge this!
- [ ] add issue to track points of improvement discussed in here
    - https://github.com/JuliaData/DataTables.jl/pull/17#discussion_r104328002 & associated generation of right_dt a few lines below it
    - https://github.com/JuliaData/DataTables.jl/pull/17#discussion_r102462335
    - https://github.com/JuliaData/DataTables.jl/pull/17#discussion_r104316856
- [ ] get https://github.com/JuliaData/DataTables.jl/pull/24 & https://github.com/JuliaData/DataTables.jl/pull/26 merged, which should make DataTables usable day to day for most people (It's working great for me!)
- [ ] add these benchmarks and develop join benchmarks
- [ ] address points of improvement (the stuff we are moving to an issue) and see if we can flip the join speeds back in favor of this change